refactor: login UI
