+++ styles/globals.css
@@ -17,3 +17,7 @@ a {
   text-decoration: none;
   color: #000;
 }
+
+input:-webkit-autofill {
+    -webkit-box-shadow: 0 0 0 1000px white inset;
+}
